Code refactored. 1) Each tab now lives on its own file 2) show_message function is not a method of the MainWindow 3) Tabs (widgets) have references to the MainWindow, so they can use the show_message function
