+++ gui/tabs/another_tab.py
@@ -1,0 +1,19 @@
+from PySide6.QtWidgets import QWidget
+
+
+class AnotherTab(QWidget):
+    def __init__(self, main_window=None):
+        super().__init__()
+        self.main_window = main_window
+        # If you want to add any widgets to this tab, you can set up a layout here and add them.
+        # For now, it's just an empty tab.
+
+        # Example of adding a layout:
+        # layout = QVBoxLayout(self)
+        # self.setLayout(layout)
+
+        # Example of adding a label to the layout:
+        # placeholder_label = QLabel("This is just a placeholder.")
+        # layout.addWidget(placeholder_label)
+
+        # You can also set up buttons, text fields, and other widgets as needed.
